Tags

+++ app/src/main/java/com/example/navsample/dto/analyzer/AnalyzedReceiptData.kt
@@ -3,8 +3,8 @@ package com.example.navsample.dto.analyzer
 data class AnalyzedReceiptData(
     var companyName: String = "",
     var valueNIP: String = "",
-    var valuePLN: Double = 0.0,
-    var valuePTU: Double = 0.0,
+    var valuePLN: Int = 0,
+    var valuePTU: Int = 0,
     var valueDate: String = "",
     var valueTime: String = ""
 )
